Add an URL

Add and URL in HTML or Markdown format.

addURL(url, text = NULL, extra = NULL, markdown = FALSE)

Arguments

url
An URL.

text
A character string to be searched within.

extra
Extra A connection object or a character string that stand for a path, ignored if text is defined.

markdown
logical. If TRUE, hyperlinks marks of Markdown language are sought.

Examples

addURL('https://github.com/inSileco/letiRmisc', 'Github repo')

#> [1] "<a href='https://github.com/inSileco/letiRmisc'>Github repo</a>"

addURL('https://github.com/inSileco/letiRmisc', 'Github repo', markdown=TRUE)

#> [1] "[Github repo](https://github.com/inSileco/letiRmisc)"