Create a data frame from scratch or based on one or two data frames.

This function handles the creation of data frames based on intuitive parameters. It was originally designed to make row binding easier when columns differs among data frame by creating data frames with the same columns.

dfTemplate(cols, nrows = 1, col_classes = "character", fill = NA)

dfTemplateMatch(x, y, yonly = FALSE, ...)

Arguments

cols
either a number of column or a vector of character used as columns names of the data frame to be returned.

nrows
row number.

col_classes
vector of column classes for the desired data frame. By default, "character" is used for all columns.

fill
character or number used to fill out the columns. Default is NA.

x
a data frame.

y
data frame or a vector of strings that includes part of or all of the column names of the data frame to be output.

yonly
a logical. Should only y (or the names(y)) be used for the data frame to be returned? Default is set to FALSE meaning that both the names of x and the names of y are used.

...
further arguments to be passed to dfTemplate.

Value

Returns a data frame with the desired characteristics.

Functions

dfTemplateMatch: Returns a data frame.

Examples

assignIds(list(2,'f', 'd', 'f'))

#> [1] 1 2 3 2