Find values in a given vector.

Given a set of values and a vector where the values must be found, the findThem function records the matched values and the position in the vector where the values have been searched for.

findThem(what, where, todf = FALSE, reportnomatch = FALSE)

Arguments

what
a vector a values to be searched for.

where
a vector where values will be searched on.

todf
a logical indicating whether the output object must be a data frame.

reportnomatch
a logical. If TRUE, values without match are reported in the data frame with a NA. Only available if todf is TRUE. Default is set to FALSE.

Value

A list indicating matched positions for each elements of what. If no match is found then NA is returned. If todf is TRUE then a three-columns data frame is returned including values and positions in both what and where vectors.

See also

Examples

x <- stats::rpois(1000,10)
findThem(c(10,4,100), x)

#> $`10`
#>   [1]   2   3   5   8  13  15  19  20  23  26  32  36  41  69  76  78  79  83
#>  [19]  86 102 105 116 117 118 119 154 159 161 163 172 180 183 192 197 198 200
#>  [37] 214 229 239 247 251 252 258 290 292 293 315 319 345 414 419 428 434 440
#>  [55] 442 466 467 477 480 488 503 515 521 532 560 561 565 567 574 579 583 594
#>  [73] 595 599 600 603 633 635 636 640 649 694 699 700 726 727 733 742 752 763
#>  [91] 764 765 769 787 790 791 806 815 819 828 833 834 835 840 852 874 879 888
#> [109] 913 920 921 922 949 965 968 972 976 980 982 983 989
#> 
#> $`4`
#> [1]  74 314 364 616 659 793
#> 
#> $`100`
#> integer(0)
#> 

findThem(c(10,4,100), x, todf=TRUE)

#>     values what where
#> 1       10    1     2
#> 2       10    1     3
#> 3       10    1     5
#> 4       10    1     8
#> 5       10    1    13
#> 6       10    1    15
#> 7       10    1    19
#> 8       10    1    20
#> 9       10    1    23
#> 10      10    1    26
#> 11      10    1    32
#> 12      10    1    36
#> 13      10    1    41
#> 14      10    1    69
#> 15      10    1    76
#> 16      10    1    78
#> 17      10    1    79
#> 18      10    1    83
#> 19      10    1    86
#> 20      10    1   102
#> 21      10    1   105
#> 22      10    1   116
#> 23      10    1   117
#> 24      10    1   118
#> 25      10    1   119
#> 26      10    1   154
#> 27      10    1   159
#> 28      10    1   161
#> 29      10    1   163
#> 30      10    1   172
#> 31      10    1   180
#> 32      10    1   183
#> 33      10    1   192
#> 34      10    1   197
#> 35      10    1   198
#> 36      10    1   200
#> 37      10    1   214
#> 38      10    1   229
#> 39      10    1   239
#> 40      10    1   247
#> 41      10    1   251
#> 42      10    1   252
#> 43      10    1   258
#> 44      10    1   290
#> 45      10    1   292
#> 46      10    1   293
#> 47      10    1   315
#> 48      10    1   319
#> 49      10    1   345
#> 50      10    1   414
#> 51      10    1   419
#> 52      10    1   428
#> 53      10    1   434
#> 54      10    1   440
#> 55      10    1   442
#> 56      10    1   466
#> 57      10    1   467
#> 58      10    1   477
#> 59      10    1   480
#> 60      10    1   488
#> 61      10    1   503
#> 62      10    1   515
#> 63      10    1   521
#> 64      10    1   532
#> 65      10    1   560
#> 66      10    1   561
#> 67      10    1   565
#> 68      10    1   567
#> 69      10    1   574
#> 70      10    1   579
#> 71      10    1   583
#> 72      10    1   594
#> 73      10    1   595
#> 74      10    1   599
#> 75      10    1   600
#> 76      10    1   603
#> 77      10    1   633
#> 78      10    1   635
#> 79      10    1   636
#> 80      10    1   640
#> 81      10    1   649
#> 82      10    1   694
#> 83      10    1   699
#> 84      10    1   700
#> 85      10    1   726
#> 86      10    1   727
#> 87      10    1   733
#> 88      10    1   742
#> 89      10    1   752
#> 90      10    1   763
#> 91      10    1   764
#> 92      10    1   765
#> 93      10    1   769
#> 94      10    1   787
#> 95      10    1   790
#> 96      10    1   791
#> 97      10    1   806
#> 98      10    1   815
#> 99      10    1   819
#> 100     10    1   828
#> 101     10    1   833
#> 102     10    1   834
#> 103     10    1   835
#> 104     10    1   840
#> 105     10    1   852
#> 106     10    1   874
#> 107     10    1   879
#> 108     10    1   888
#> 109     10    1   913
#> 110     10    1   920
#> 111     10    1   921
#> 112     10    1   922
#> 113     10    1   949
#> 114     10    1   965
#> 115     10    1   968
#> 116     10    1   972
#> 117     10    1   976
#> 118     10    1   980
#> 119     10    1   982
#> 120     10    1   983
#> 121     10    1   989
#> 122      4    2    74
#> 123      4    2   314
#> 124      4    2   364
#> 125      4    2   616
#> 126      4    2   659
#> 127      4    2   793

findThem(c(10,4,100), x, todf=TRUE, reportnomatch=TRUE)

#>     values what where
#> 1       10    1     2
#> 2       10    1     3
#> 3       10    1     5
#> 4       10    1     8
#> 5       10    1    13
#> 6       10    1    15
#> 7       10    1    19
#> 8       10    1    20
#> 9       10    1    23
#> 10      10    1    26
#> 11      10    1    32
#> 12      10    1    36
#> 13      10    1    41
#> 14      10    1    69
#> 15      10    1    76
#> 16      10    1    78
#> 17      10    1    79
#> 18      10    1    83
#> 19      10    1    86
#> 20      10    1   102
#> 21      10    1   105
#> 22      10    1   116
#> 23      10    1   117
#> 24      10    1   118
#> 25      10    1   119
#> 26      10    1   154
#> 27      10    1   159
#> 28      10    1   161
#> 29      10    1   163
#> 30      10    1   172
#> 31      10    1   180
#> 32      10    1   183
#> 33      10    1   192
#> 34      10    1   197
#> 35      10    1   198
#> 36      10    1   200
#> 37      10    1   214
#> 38      10    1   229
#> 39      10    1   239
#> 40      10    1   247
#> 41      10    1   251
#> 42      10    1   252
#> 43      10    1   258
#> 44      10    1   290
#> 45      10    1   292
#> 46      10    1   293
#> 47      10    1   315
#> 48      10    1   319
#> 49      10    1   345
#> 50      10    1   414
#> 51      10    1   419
#> 52      10    1   428
#> 53      10    1   434
#> 54      10    1   440
#> 55      10    1   442
#> 56      10    1   466
#> 57      10    1   467
#> 58      10    1   477
#> 59      10    1   480
#> 60      10    1   488
#> 61      10    1   503
#> 62      10    1   515
#> 63      10    1   521
#> 64      10    1   532
#> 65      10    1   560
#> 66      10    1   561
#> 67      10    1   565
#> 68      10    1   567
#> 69      10    1   574
#> 70      10    1   579
#> 71      10    1   583
#> 72      10    1   594
#> 73      10    1   595
#> 74      10    1   599
#> 75      10    1   600
#> 76      10    1   603
#> 77      10    1   633
#> 78      10    1   635
#> 79      10    1   636
#> 80      10    1   640
#> 81      10    1   649
#> 82      10    1   694
#> 83      10    1   699
#> 84      10    1   700
#> 85      10    1   726
#> 86      10    1   727
#> 87      10    1   733
#> 88      10    1   742
#> 89      10    1   752
#> 90      10    1   763
#> 91      10    1   764
#> 92      10    1   765
#> 93      10    1   769
#> 94      10    1   787
#> 95      10    1   790
#> 96      10    1   791
#> 97      10    1   806
#> 98      10    1   815
#> 99      10    1   819
#> 100     10    1   828
#> 101     10    1   833
#> 102     10    1   834
#> 103     10    1   835
#> 104     10    1   840
#> 105     10    1   852
#> 106     10    1   874
#> 107     10    1   879
#> 108     10    1   888
#> 109     10    1   913
#> 110     10    1   920
#> 111     10    1   921
#> 112     10    1   922
#> 113     10    1   949
#> 114     10    1   965
#> 115     10    1   968
#> 116     10    1   972
#> 117     10    1   976
#> 118     10    1   980
#> 119     10    1   982
#> 120     10    1   983
#> 121     10    1   989
#> 122      4    2    74
#> 123      4    2   314
#> 124      4    2   364
#> 125      4    2   616
#> 126      4    2   659
#> 127      4    2   793
#> 128    100    3    NA