Keep these words.

Extract words based on their position in a character string, vectorised over string.

keepWords(string, slc = 1, punct.rm = TRUE, na.rm = FALSE,
  collapse = NULL)

Arguments

string
Input vector. A character vector or list from which words will be extracted.

slc
Vector of integer indicating the selected positions of the words to be kept.

punct.rm
A logical indicating whether punctuation characters should be removed.

na.rm
A logical indicating whether missing values should be removed.

collapse
An optional character string used to separate selected words. Not NA_character_.

Value

A vector of the selected words concatenated for each string.

Examples

strex <- 'Lorem ipsum dolor sit amet, consectetur adipisicing elit.'
keepWords(strex)

#> [[1]]
#> [1] "Lorem"
#> 

keepWords(c(strex,'A second chacter string.'), slc=c(1,8), na.rm=TRUE)

#> [[1]]
#> [1] "Lorem" "elit" 
#> 
#> [[2]]
#> [1] "A"
#> 

keepWords(c(strex,'A second chacter string.'), slc=c(1,8), na.rm=TRUE, collapse='/')

#> [[1]]
#> [1] "Lorem/elit"
#> 
#> [[2]]
#> [1] "A"
#>