Scale a set of values

For a given set of values, returns values between 1 and a number of values passed as an argument.

scaleWithin(x, nval = 100, mn = min(x), mx = max(x))

Arguments

x
a numeric object of type numeric or coercible as one.

nval
The number of scaled values (between 1 and nval).

mn
Minimum value (mn and lower values are set to 1).

mx
Maximum value (mx and higher values are set to nval).

Value

A numeric vector

Examples

x <- stats::rpois(20, 20)
scaleWithin(x, 20, 10, 30)

#>  [1] 10 13 13 12  9 12 15 14  5  9 13 15  5  9 12  9  5  6  9 13

scaleWithin(matrix(x, 10, 2), mn = 20, mx = 30)

#>       [,1] [,2]
#>  [1,]    1   21
#>  [2,]   21   41
#>  [3,]   21    1
#>  [4,]   11    1
#>  [5,]    1   11
#>  [6,]   11    1
#>  [7,]   41    1
#>  [8,]   31    1
#>  [9,]    1    1
#> [10,]    1   21